Change IE bug tracker link to archived version
As of January 3rd, 2018, Microsoft Connect is no longer in service [1]
and as such IE bug tracker links don't work anymore. It doesn't seem
like the tracker was migrated anywhere; it was just shut down
(probably because they aren't fixing anything but the most critical
bugs).

Although its usefulness is limited because the bug wont be fixed, I
think the bug tracking page may still be good for people to see this
fact. I've updated the link to the latest archived version from the
Wayback Machine.

[1]: https://docs.microsoft.com/en-us/collaborate/connect-redirect